Add language and max players setting interface
Add timestamp to blocked listings. This should be updated each time we see the listing, so we can prune blocked listings we haven't seen for a while (maybe 30 days).
Also lay some groundwork for changing blockUsers from an array of user IDs to a mapping from user ID to username so we can better manage blocked users.